backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic>=2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1

# AI and NLP (Optimized)
google-generativeai==0.8.3
transformers==4.36.0
torch==2.2.0
sentence-transformers>=2.5.0
spacy>=3.8.0
nltk==3.8.1
textstat==0.7.3

# Document processing
pypdf2==3.0.1
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3

# Data processing
pandas==2.1.4
numpy>=1.26.0
scikit-learn==1.3.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
